extract event date helpers
