+++ app/src/helpers/dateHelper.ts
@@ -1,0 +1,3 @@
+export function toUnixSeconds(value: Date): number {
+    return Math.floor(value.getTime() / 1000);
+}
