+++ app/src/models/event/Event.ts
@@ -2,12 +2,14 @@ import mongoose, { HydratedDocument, Schema, Types, SchemaTypes, Model } from "m
 import { type Filterable } from "@/types/Filter";
 import { type CurrencyCode } from "currency-codes-ts/dist/types";
 import { getConfigValue } from "@/helpers/configHelper";
+import { toUnixSeconds } from "@/helpers/dateHelper";
 import { type Sortable } from "@/types/Sort";
 import { paginatePlugin, type PaginateQueryHelper } from "@/models/plugins/paginate";
 import messages from "@/constants/errorMessages";
 import { getFilterableFields, getSortableFields } from "@/models/event/statics";
 import { categoriesValidator, typeValidator } from "@/models/event/validators";
 import { postSaveHook, preSaveHook } from "@/models/event/hooks";
+import { normalizeEventDate } from "@/models/event/utils";
 
 export interface IEvent {
     _id: Types.ObjectId;
@@ -35,37 +37,6 @@ export interface IEvent {
 export type HydratedEvent = HydratedDocument<IEvent>;
 
 interface EventModelType extends Model<IEvent, PaginateQueryHelper<IEvent>, object>, Filterable, Sortable { }
-
-function normalizeEventDate(value: Date | number | string): Date {
-    if (value instanceof Date) {
-        return value;
-    }
-
-    if (typeof value === "number") {
-        const milliseconds = Math.abs(value) >= 1e12 ? value : value * 1000;
-        return new Date(milliseconds);
-    }
-
-    if (typeof value === "string") {
-        const trimmedValue = value.trim();
-
-        if (/^-?\d+$/.test(trimmedValue)) {
-            return normalizeEventDate(Number(trimmedValue));
-        }
-
-        return new Date(trimmedValue);
-    }
-
-    return value;
-}
-
-function toUnixSeconds(value: unknown): number | unknown {
-    if (!(value instanceof Date)) {
-        return value;
-    }
-
-    return Math.floor(value.getTime() / 1000);
-}
 
 export const EventSchema = new Schema<IEvent, Model<IEvent>, object, PaginateQueryHelper<IEvent>>(
     {
@@ -169,7 +140,9 @@ EventSchema.set('toJSON', {
             delete transformedRet.updatedAt;
         }
 
-        transformedRet.date = toUnixSeconds(transformedRet.date);
+        if (transformedRet.date instanceof Date) {
+            transformedRet.date = toUnixSeconds(transformedRet.date);
+        }
         transformedRet.imageUrl = getConfigValue('HOST') + '/' + transformedRet.imageUrl
 
         return transformedRet;
